+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * added option to return `UpdateEvent::Capture` from the `update()` function to capture in-game screenshots
 * added `max_fps` to `AppOptions` to lower the frame rate even below the screen vsync frequency. See lowfps example.
 * switched to unrust 0.2 to run on the latest winit/glutin/wasm_bindgen/web-sys versions
+* upgraded to image 0.23.14 to fix a vulnerability
 
 # 1.2.3 - 2020 Jan 08
 ## fixed